Split credential registry property tests
Move SharedCredentialRegistry property and concurrency coverage into a
sibling WASM test module so `credential_injector_tests.rs` stays below
the repository line-count limit while preserving the existing test names.
